requirements.txt
pybind11
fastapi
uvicorn
absl-py==1.4.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
einops==0.6.0
filelock==3.9.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.12.1
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
joblib==1.2.0
keopscore==2.1.1
kiwisolver==1.4.4
librosa==0.9.2
lightning-utilities==0.6.0.post0
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pykeops==2.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.97
six==1.16.0
SoundFile==0.10.3.post1
speechbrain==0.5.13
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torchmetrics==0.6.0
tqdm==4.64.1
transformers==4.26.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.13.0
setuptools==59.5.0